Fixed spoon and kid animatin
Spoon animation fixed:
the spoon rotates when it reached near of the kid and the animation is palyed between certain distance between the kid and the food.

Kid Animation fixed:
When the spoon becomes full ( there is food on the spoon ) the kid opens his mouth and when the spoon touches the kid then the kid closes his mouth.