Reception window dialogue and reception room
Things done:
We have 2 scenes. When the user clicks on the reception window in the reception room scene, the second scene is displayed where a nurse is displayed welcoming the user and instructing him of what do next ( E.g. click on the ingestion room to get food ).

Possible improvement:
The dialogue becomes dynamic depending on the game settings that has been chosen from the beginning of the game ( ingestion/injection and age ).