Added the bubble game
made 3 scripts: Bubble, BubbleClick,  HideBottleBg

the first 2 second an animaiton for the bottle background will be played. Then the bottle background will be hidden and bubbles will appear. The blue bubble needs only one click to be popped while the pink one needs 2 clicks to be popped.

The script is made so that it is possible to add aditional bubbles of all the previous 2 colours as well as green bubbles which should be popped after 3 clicks. The game now is scalled down so that other members can test, evaluate and give additional ideas.

The game ends when 3 bubbles are dispalyed on the screen at the same time, so the player needs to pop bubbles quickly. The player will win when they finish the milk which they can click on while bubbles appearing or after popping all the bubbles.