Added movement & changed size of pink bubbles
Things have been done:

Added movement for each bubble ( using code and not the animation feature ). So that each bubble moves smoothly toward the baby bottle.

Changed the scale of the pink bubbles so that it is almost twice the size of the blue bubbles.

possible features to be added:
popping up animation when a bubble is clicked.