Added arrow to first bubble
As suggested by Amin, the first bubble has now an arrow that will notify the player that he/she needs to click on it. The time starts at the start of the game and is paused after first bubble is shown then the time runs again once the first bubble is popped.

The movement of the bubble needs a bit of improvement after pausing the time as the movement of the bubbles continues running in the background.