partly fixed movement problem
fixed the problem that I faced when time was paused. But I have just realised that all the bubbles are moving at the same time after the first bubble is popped. So we need the movement to start only when the bubbles appear as there is movement when they are hidden and that is something that needs to be fixed.