Added arrows
I added arrow that notified the user that they need to click on a blue buble and then another arrow for a pink bubble. The pink bubble has x1 and when clicked x2 is displayed indicating that the pink bubbles need 2 clicks. the timing is paused when each arrow is dispalyed. The movement of other bubbles still needs fixing so that it will start when the bubble appear and not when the second buble is popped ( the first pink one )