Working On Tutorial For the matching game
So now the first card is animated in a way to inform the player that its needs to be clicked. And when it's being clicked we will get the index of the card with similar image.  Note that all the cards will be disabled so that the user must click on the card that has the animation.

To do:
animate a second card ( after pressing on the first card ) so that it will not have same image ( not a match ) then a losing sound will be played to inform the user that he/she didn't get it right.
Then a third card will be animated ( after 1st and 2nd being pressed ) and it will be a match then when pressed a wining sound will be played.