Added sound & one wrong card and one correct card
Now when the player pressed on the first card they will be instructred to presses on a second card which is not a match and when the player presses on it then a sad trombone sound will be played. After that the user should press on the first card again ( this will be done soon ) and then press on the second animated card ( the correct match ) and when presses a cheering sound will be palyed.