Fixed 2 bugs, refacotred and commented the code
fixed the bug so that now when the cards are facing up then the card will be disabled ( it will become unclickable ) and when faced back then it will be enabled again. and the card that have been matched will be disabled as well as those two bugs caused problem for the tutorial. unused code is deleted and the whole classes are commented out so that it is ready to be tested.